requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# ML and NLP
sentence-transformers==2.4.0
huggingface_hub==0.20.3
chromadb==0.4.18
datasets==2.15.0
numpy==1.24.3

# Google Gemini
google-generativeai==0.3.2

# Optional frontend
streamlit==1.28.1

# Utilities
loguru==0.7.2
httpx==0.25.2
pytest==7.4.3
pydantic-settings==2.1.0
langchain==0.1.16

